+++ inst/htmlwidgets/calendar.js
@@ -313,7 +313,6 @@ if (HTMLWidgets.shinyMode) {
   Shiny.addCustomMessageHandler("proxy-toastui-calendar-delete", function(obj) {
     var cal = get_widget(obj.id);
     if (typeof cal != "undefined") {
-      console.log(obj.data);
       var scheduleId = obj.data.scheduleId;
       var calendarId = obj.data.calendarId;
       for (let i = 0; i < scheduleId.length; i += 1) {
@@ -339,6 +338,15 @@ if (HTMLWidgets.shinyMode) {
     var cal = get_widget(obj.id);
     if (typeof cal != "undefined") {
       cal.setOptions(obj.data.options);
+    }
+  });
+  Shiny.addCustomMessageHandler("proxy-toastui-calendar-toggle", function(obj) {
+    var cal = get_widget(obj.id);
+    if (typeof cal != "undefined") {
+      var calendarId = obj.data.calendarId;
+      for (let i = 0; i < calendarId.length; i += 1) {
+        cal.toggleSchedules(calendarId[i], obj.data.toHide);
+      }
     }
   });
 }